src/main/java/cn/xpbootcamp/gilded_rose/Product.java
package cn.xpbootcamp.gilded_rose;

public class Product {
    private int sellIn;
    private int quality;

    public Product(Integer sellIn, Integer quality) {
        this.sellIn = sellIn;
        this.quality = quality;
    }

    public int getSellIn() {
        return sellIn;
    }

    public int getQuality() {
        return quality;
    }

    public void setSellIn(int sellIn) {
        this.sellIn = sellIn;
    }

    public void setQuality(int quality) {
        this.quality = quality;
    }
}
